reverting to float

+++ structures_typedef/dog.h
@@ -13,7 +13,7 @@
 typedef struct dog
 {
 char *name;
-int age;
+float age;
 char *owner;
 } dog_t;
